nint/nuint enum conversion

+++ src/AVKit/Enums.cs
@@ -6,7 +6,7 @@ namespace AVKit {
 #if !XAMCORE_3_0 || MONOMAC
 	// this enum only exists for OSX (not iOS)
 	[Native]
-	public enum AVPlayerViewControlsStyle : nint {
+	public enum AVPlayerViewControlsStyle : long {
 		None,
 		Inline,
 		Floating,
@@ -18,7 +18,7 @@ namespace AVKit {
 #if MONOMAC
 	[Mac (10,10)]
 	[Native]
-	public enum AVCaptureViewControlsStyle : nint {
+	public enum AVCaptureViewControlsStyle : long {
 		Inline,
 		Floating,
 		InlineDeviceSelection,
@@ -27,7 +27,7 @@ namespace AVKit {
 
 	[Mac (10,9)]
 	[Native]
-	public enum AVPlayerViewTrimResult : nint {
+	public enum AVPlayerViewTrimResult : long {
 		OKButton,
 		CancelButton
 	}
@@ -37,7 +37,7 @@ namespace AVKit {
 	[iOS (9,0)]
 	[Native]
 	[ErrorDomain ("AVKitErrorDomain")]
-	public enum AVKitError : nint {
+	public enum AVKitError : long {
 		None = 0,
 		Unknown = -1000,
 		PictureInPictureStartFailed = -1001
