+++ src/UIKit/UIGuidedAccessRestriction.cs
@@ -21,7 +21,7 @@ namespace UIKit {
 	// NSInteger -> UIGuidedAccessRestrictions.h
 	[Native]
 	[iOS (7,0)]
-	public enum UIGuidedAccessRestrictionState : nint {
+	public enum UIGuidedAccessRestrictionState : long {
 		Allow,
 		Deny
 	}
